Rename examplemod identifiers to kubejsgui
