+++ src/main/java/com/example/kubejsgui/menu/RecipeEditorMenu.java
@@ -1,6 +1,6 @@
-package com.example.examplemod.menu;
+package com.example.kubejsgui.menu;
 
-import com.example.examplemod.registry.ModMenuTypes;
+import com.example.kubejsgui.registry.ModMenuTypes;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
