--- src/main/resources/assets/examplemod/lang/en_us.json
@@ -1,7 +1,0 @@
-{
-  "itemGroup.examplemod": "Example Mod Tab",
-  "block.examplemod.example_block": "Example Block",
-  "item.examplemod.example_item": "Example Item",
-  "key.categories.examplemod": "Example Mod",
-  "key.examplemod.open_gui": "Open Blank GUI"
-}
