+++ src/main/templates/META-INF/mods.toml
@@ -38,7 +38,7 @@ displayName="${mod_name}" #mandatory
 #displayURL="https://change.me.to.your.mods.homepage.example.invalid/" #optional
 
 # A file name (in the root of the mod JAR) containing a logo for display
-#logoFile="examplemod.png" #optional
+#logoFile="kubejsgui.png" #optional
 
 # A text field displayed in the mod UI
 #credits="" #optional
